Вынес Демки в отдельный каталог
