--- frontend/src/pages/AddChartPage.js
@@ -1,95 +1,0 @@
-import React from 'react';
-import { Grid, makeStyles, Paper, Typography } from '@material-ui/core';
-import { ResponsiveContainer, LineChart, Line, XAxis, YAxis, CartesianGrid, Tooltip, Legend } from 'recharts';
-
-const data = [
-  {
-    name: 'Page A',
-    uv: 4000,
-    pv: 2400,
-    amt: 2400,
-  },
-  {
-    name: 'Page B',
-    uv: 3000,
-    pv: 1398,
-    amt: 2210,
-  },
-  {
-    name: 'Page C',
-    uv: 2000,
-    pv: 9800,
-    amt: 2290,
-  },
-  {
-    name: 'Page D',
-    uv: 2780,
-    pv: 3908,
-    amt: 2000,
-  },
-  {
-    name: 'Page E',
-    uv: 1890,
-    pv: 4800,
-    amt: 2181,
-  },
-  {
-    name: 'Page F',
-    uv: 2390,
-    pv: 3800,
-    amt: 2500,
-  },
-  {
-    name: 'Page G',
-    uv: 3490,
-    pv: 4300,
-    amt: 2100,
-  },
-];
-
-const useStyles = makeStyles(theme => ({
-  chartPaper: {
-    padding: theme.spacing(2),
-  },
-  chartHeader: {
-    color: theme.palette.primary.main,
-    fontSize: 18,
-  },
-}));
-
-const AddChartPage = () => {
-  const styles = useStyles();
-
-  return (
-    <Grid container xs={12}>
-      <Grid item xs={12} md={6}>
-        <Paper className={styles.chartPaper}>
-          <Typography align="center" color="textPrimary" gutterBottom component="h5" className={styles.chartHeader}>
-            Line chart
-          </Typography>
-          <ResponsiveContainer width="100%" height={300}>
-            <LineChart
-              data={data}
-              margin={{
-                top: 5,
-                right: 30,
-                left: 20,
-                bottom: 5,
-              }}
-            >
-              <CartesianGrid strokeDasharray="3 3" />
-              <XAxis dataKey="name" />
-              <YAxis />
-              <Tooltip />
-              <Legend />
-              <Line type="monotone" dataKey="pv" stroke="#8884d8" activeDot={{ r: 8 }} />
-              <Line type="monotone" dataKey="uv" stroke="#82ca9d" />
-            </LineChart>
-          </ResponsiveContainer>
-        </Paper>
-      </Grid>
-    </Grid>
-  );
-};
-
-export default AddChartPage;
