+++ frontend/src/pages/AddChartPage/ChartDemos/demosMap.js
@@ -1,0 +1,3 @@
+import LinearChartDemo from './LinearChartDemo';
+
+export default [{ title: 'Line chart', component: LinearChartDemo }];
