+++ frontend/src/pages/AddChartPage/ChartDemos/index.js
@@ -1,0 +1,1 @@
+export { default } from './demosMap';
