+++ frontend/src/pages/AddChartPage/index.js
@@ -1,0 +1,1 @@
+export { default as AddChartPage } from './AddChartPage';
