+++ frontend/src/pages/index.js
@@ -1,2 +1,2 @@
 export { default as MainPage } from './MainPage';
-export { default as AddChartPage } from './AddChartPage';
+export { AddChartPage } from './AddChartPage';
